addressed the nits
